+++ src/test/java/io/supertokens/test/TestingProcessManager.java
@@ -20,8 +20,14 @@ import io.supertokens.Main;
 import io.supertokens.ProcessState;
 import io.supertokens.ProcessState.EventAndException;
 import io.supertokens.ProcessState.PROCESS_STATE;
+import io.supertokens.pluginInterface.STORAGE_TYPE;
+import io.supertokens.storageLayer.StorageLayer;
 
 import java.util.ArrayList;
+import java.util.function.Consumer;
+import java.util.function.Predicate;
+
+import static org.junit.Assert.assertNotNull;
 
 public class TestingProcessManager {
 
@@ -144,5 +150,20 @@ public class TestingProcessManager {
             }
             return e;
         }
+    }
+
+    /**
+     * Utility function to wrap tests with, as they require TestingProcess
+     */
+    public static void withProcess(ProcessConsumer consumer) throws Exception {
+        String[] args = {"../"};
+
+        TestingProcessManager.TestingProcess process = TestingProcessManager.start(args);
+        assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STARTED));
+
+        consumer.accept(process);
+
+        process.kill();
+        assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STOPPED));
     }
 }